refactor: move components and hooks from maps tab folder
